feat(views): add API controller and views
Add API v1 controller with standard views from old version
